pyproject.toml
[tool.poetry]
name = "agent47"
version = "0.1.0"
description = "Agent 47: A lethal multi-agent system for autonomous bug fixing and code refactoring in isolated environments."
authors = ["Faithful King <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = "1.2.10"
langgraph = "1.0.10"
langchain-google-genai = "4.2.1"
docker = "7.1.0"
python-dotenv = "1.2.2"
pydantic = "2.12.5"
annotated-doc = "0.0.4"
annotated-types = "0.7.0"
anthropic = "0.84.0"
anyio = "4.12.1"
bracex = "2.6"
certifi = "2026.2.25"
cffi = "2.0.0"
charset-normalizer = "3.4.4"
click = "8.3.1"
colorama = "0.4.6"
cryptography = "46.0.5"
deepagents = "0.4.4"
distro = "1.9.0"
docstring-parser = "0.17.0"
fastapi = "0.135.1"
filetype = "1.2.0"
gitdb = "4.0.12"
gitpython = "3.1.46"
google-auth = "2.48.0"
google-genai = "1.65.0"
greenlet = "3.3.2"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
idna = "3.11"
jiter = "0.13.0"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
langchain-anthropic = "1.3.4"
langchain-core = "1.2.16"
langgraph-checkpoint = "4.0.1"
langgraph-prebuilt = "1.0.8"
langgraph-sdk = "0.3.9"
langsmith = "0.7.9"
orjson = "3.11.7"
ormsgpack = "1.12.2"
packaging = "26.0"
psycopg2-binary = "2.9.11"
pyasn1 = "0.6.2"
pyasn1-modules = "0.4.2"
pycparser = "3.0"
pydantic-core = "2.41.5"
pygithub = "2.8.1"
pyjwt = "2.11.0"
pynacl = "1.6.2"
pywin32 = "311"
pyyaml = "6.0.3"
requests = "2.32.5"
requests-toolbelt = "1.0.0"
rsa = "4.9.1"
smmap = "5.0.2"
sniffio = "1.3.1"
sqlalchemy = "2.0.48"
starlette = "0.52.1"
tenacity = "9.1.4"
typing-inspection = "0.4.2"
typing-extensions = "4.15.0"
urllib3 = "2.6.3"
uuid-utils = "0.14.1"
uvicorn = "0.41.0"
wcmatch = "10.1"
websockets = "16.0"
xxhash = "3.6.0"
zstandard = "0.25.0"
celery = "^5.6.2"
redis = "^7.3.0"
langchain-huggingface = "^1.2.1"
huggingface-hub = "^1.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
